Add files via upload
This is the final update for connectivity. the BOC web-app now has all functional web services as well as servlets to handle the communication between the jsp and the web service clients. included is the example robin gave me of how to use servlets to process requests. any code that needs to be tested should be done on the server. the folder where everything is located is CPTR333